requirements.txt
numpy>=1.24
scipy>=1.10
casadi>=3.6
matplotlib>=3.7
opencv-python>=4.8
PyYAML>=6.0
tqdm>=4.66
